feat: upgrade go dependencies
This commit upgrades the go dependencies to their latest versions.
The gosec linter was disabled for some lines because they are not
security issues.
The expire package was removed because it was not being used.
The expire_test package was moved to beacon_test package.
The CalculateSlotExpiration and GetSlotTime functions were moved to
the expire_test package.

+++ pkg/beacon/download.go
@@ -172,8 +172,9 @@ func (d *Default) fetchHistoricalCheckpoints(ctx context.Context, checkpoint *v1
 	// Calculate the epoch boundaries we need to fetch
 	// We'll derive the current finalized slot and then work back in intervals of SLOTS_PER_EPOCH.
 	currentSlot := uint64(checkpoint.Finalized.Epoch) * uint64(sp.SlotsPerEpoch)
-	for i := uint64(1); i < uint64(d.config.HistoricalEpochCount); i++ {
-		slot := phase0.Slot(currentSlot - i*uint64(sp.SlotsPerEpoch))
+	for i := 1; i < d.config.HistoricalEpochCount; i++ {
+		//nolint:gosec // This is not a security issue
+		slot := phase0.Slot(currentSlot - uint64(i)*uint64(sp.SlotsPerEpoch))
 		slotsInScope[slot] = struct{}{}
 	}
